Squashed commit of the following:
commit 1043c1b8b3bd2f8e14a8f0049cfba0899e38453d
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 26 12:12:26 2013 +0900

    Delete trailing white spaces and convert tab to four spaces

commit 005df4a0bbf4ec4c0f0dffc72db4ad820a70ad36
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Mon Mar 25 18:04:05 2013 +0900

    Format source codes

commit a541ac19beddad5adf56979420b9a3316dfeefdf
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Fri Mar 15 10:44:43 2013 +0900

    Modify pom.xml

commit 18cb7abd7c9452923df6439d6d4c85cf953932d8
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Wed Mar 13 16:32:04 2013 +0900

    Modify gitignore

commit 7cd4b9187e8d7da2a952653a72e6203d7d56796d
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Wed Mar 13 16:31:43 2013 +0900

    Remove Eclipse invisible files

commit 9dbbea3891c4152cef8c7a3be56f49d8db245168
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 12 05:05:07 2013 +0900

    Fix readme

commit dc44a7adc081039ea8e345d86a588af39c9c096f
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 12 04:32:12 2013 +0900

    Remove sound package

commit d02eb1fba6e695f1cd751bdd654183f2e17036c2
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 5 18:14:14 2013 +0900

    Remove Cron class

commit bc7d506057bac11a7dad25885c91302afbdc8739
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 5 17:03:33 2013 +0900

    Modify to use jogl v2.0-rc11 at maven central repository

commit e8eafbb10fcb61571ce69e61aa6013c74e708908
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Mar 5 16:09:02 2013 +0900

    Modify pom.xml to use maven central repository for jsonic

commit 7706e2c25ab5583990bbad3e5a7e88a56d817d3f
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Thu Feb 14 19:05:12 2013 +0900

    Fix a bug about setFPS

commit 3460d1f6a66394448a697c1d19ae54debfcb2f3b
Merge: 5856c52 d353026
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Thu Feb 14 18:55:16 2013 +0900

    Merge branch 'dev' of github.com:casmi/casmi into dev

commit d353026653e405b32dede7c83f793d56ee4be8d5
Merge: 2efa717 a9e62ae
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Wed Jan 9 04:03:31 2013 +0900

    Merge branch 'dev' of github.com:casmi/casmi into dev

commit 2efa717d6e9f2db585009b6f444a2d4c936d4b2c
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Wed Jan 9 04:02:40 2013 +0900

    add dashedLine function

commit 5856c522a0c623e2fdc2770a00cb7d7a2e29a486
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Fri Dec 28 17:52:10 2012 +0900

    Add UI classes

commit a9e62ae427689204dc056b0d4d1bdb13a2db6346
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Fri Dec 28 16:41:52 2012 +0900

    Update dev version to 0.3.2-SNAPSHOT

commit bc2ab71fb37e3c81144a9426b7ab43672c4384f7
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Fri Dec 28 15:26:03 2012 +0900

    Modify pom.xml

commit 2f5c9561124968474493c1396853595a978e92d4
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Tue Dec 18 17:05:44 2012 +0900

    Add GraphicsUtil class

commit 9b7ffaee69a9fcdbcc413b19dbb541680673679a
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Thu Dec 6 16:25:20 2012 +0900

    Rename Graph package to Chart

commit 01307f1d25acad993dc087a3019344dbd6a59f97
Author: Yuki Ban <yuki@xcoo.jp>
Date:   Thu Dec 6 16:24:54 2012 +0900

    Rename Graph package to Chart.

commit e5cdc28fe6150d51966ddc18ac7f77c73e1f2eff
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Thu Nov 15 18:30:42 2012 +0900

    Modify JUnit dependency

commit 5741e7478eb267a5a93a41bed8ec81ad9bbf7333
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Thu Nov 15 18:30:03 2012 +0900

    Add setVisible method to Element

    Add setVisible method to Element.
    Make visible and hidden method deprecated.

commit 273950f5f4c01374b4a4451964f3a43d763d48a6
Merge: 5187780 806547b
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Sat Nov 10 00:01:40 2012 +0900

    Merge remote-tracking branch 'origin/master' into dev

    Conflicts:
    	pom.xml

commit 518778013fc7a80443594050a52bcca2e1e04f3d
Author: Toshiki TAKEUCHI <take@xcoo.jp>
Date:   Fri Nov 9 14:29:55 2012 +0900

    Update version to 0.3.1-SNAPSHOT

+++ src/main/java/casmi/chart/view/Axis.java
@@ -31,7 +31,7 @@ import casmi.graphics.group.Group;
 
 /**
  * Axis Class for Chart.
- * 
+ *
  * @author Y.Ban
  */
 public class Axis extends Group {
@@ -165,7 +165,7 @@ public class Axis extends Group {
             name.setAlign(TextAlign.CENTER);
             name.setStrokeColor(ColorSet.WHITE);
             add(name);
-            double diff = length / (double)(datas.getSize() / divisionDiff);
+            double diff = length / (datas.getSize() / divisionDiff);
             for (int i = 0; i * divisionDiff < datas.getSize(); i++) {
                 Line l = new Line(i * diff + offset, 0, i * diff + offset, -10);
                 l.setStrokeColor(ColorSet.WHITE);
@@ -184,7 +184,7 @@ public class Axis extends Group {
             name.setRotation(90);
             name.setStrokeColor(ColorSet.WHITE);
             add(name);
-            double diff = length / (double)((maxValue - minValue) / divisionDiff);
+            double diff = length / ((maxValue - minValue) / divisionDiff);
             for (int i = 0; this.minValue + i * divisionDiff <= this.maxValue; i++) {
                 Line l = new Line(0, i * diff + offset, -10, i * diff + offset);
                 l.setStrokeColor(ColorSet.WHITE);
